20110594_Nguyễn Minh Tuấn_17110043_Bùi Văn Lâm_DoAnTriTueNhanTao_NhanDienChuSoVietTay/NhanDangChuSoVietTay/requirements.txt
tensorflow
opencv-python
nump
keras
Flask